requirements.txt
pyTelegramBotAPI==4.26.0
python-decouple==3.8
requests==2.32.3